added provider detail dialog, to be improved
